"Shown" is not inclusive enough
We all know that "shown" means presented, but it is important that we include / permit user agents and users that might not "see" things.  It is also important that we permit use models like a user agent in an automobile that might be presenting the offer to the user audibly.