Remove note and add description for diagram
The diagram as it stood was not accessible.  Also, it was in a "note", which was not really needed and made it format poorly in a printed version.  Since notes are non-normative by definition, also marked the section as informative so that it is tagged accordingly.

This should really have a more accessible description if it is important.  The right way to do that is to mark up the SVG.  I have not done that at this time.