Proposal for dealing with specification of required fields
I'm suggesting a sequence string field that allows mandatory fields to
be specified and have removed some of the 'required' specifications in
the BasicCardResponse, related to discussion on
https://github.com/w3c/browser-payment-api/pull/9. I've only left PAN as
required as I think that is a safe assumption, then the merchant can
specify for example that CSC must be supplied for a particular
transaction using the requiredFields sequence in the BaseRequestData
dictionary.

I'm suggesting that if we like this model, we would move this
BaseRequestData dictionary object into the PaymentRequest API spec to
allow other payment methods to extend it.

I'm sure there are other ways to do this and am a newbie to
WebIDL/JavaScript APIs, so please forgive  if I've modelled this
completely wrong for the language. I'll learn!