Update complete() method to take a string and clarify its purpose
The argument to complete is used to influence the user experience as the
user interface for the payment request is dismissed. Some sites will not
be able to provide an indication of success (perhaps because they
require additional flow steps after the payment request before accepting
an order). Calling complete() with no arguments uses the default value
where success or failure are not indicated.

NOTE: the current WebIDL parser used by Respec does not handle default
values for enums though these are allowed in WebIDL. This results in a
Respect error once this change is merged.