Updated examples to reflect changes in recent PRs
PRs #158, #160, and #161 made changes that affect the examples. These
weren't reflected in original PRs to keep those changes simple and to
avoid merge conflicts from changing the same lines in the spec. This
changes updates the examples.

Also updating the syntax to be consistent and use JavaScript (rather
than JSON).

Fixes #164.