Adding support for phone and email
This is a proposed solution for issue #1 and a counter proposal to PR
#65.

This change adds support for two new `PaymentOptions` for payer's email
and payer's phone number. These values are provided on the
`PaymentResponse` object. Since these values should not change the total
price, no `PaymentRequestUpdate` event is fired for them.

The change also adds a phone number to the `ShippingAddress` data
structure with the implication that this gets populated when
`requestShipping` is `true`.

Here are some variations to this change that should be considered:
* Do not include `requestPayerPhone` (issue #1 actually only mentions
shipping phone number)
* Do include a `requestShippingPhone` value and only populate the
`ShippingAddress.phone` value if both `requestShipping` and
`requestShippingPhone` are true.