Remove delegated state and make abort() return a promise
The current spec defines a `delegated` state but there is no way to
observe a change into this state. The state concept is used to help
describe algorithms in the spec in a succinct way and isn't exposed to
web pages.

This change removes the delegated state from the spec. Previously the
only impact of delegated was to prevent abort() from being called.
Determining whether abort is possible might require asynchronous
processing. For example, some implementations may support abort even if
they have delegated to a payment app but may not be able to tell
synchronously if this is possible.

With this change, the abort() method attempts to abort the request and
returns a promise while it tries. If it is not possible to abort then
the promise is rejected (instead of throwing). If abort completes
successfully then the promise is resolved.