Add error message that can be provided as part of updateWith (#257)
We have had lots of feedback that merchants need to explain errors to
customers during the payment request. For example, if a shipping address
is not acceptable then the site must be able to indicate why.

This change adds an `error` string to the `PaymentDetails`, which may
only be used as part of `updateWith`.