Remove restriction that payment method identifier only appear once (#261)
Based on recent discussions about how to support payment modifiers with
basic-card and different card types, we agreed that we need to support
modifiers with the same identifier but different data. This change
removes the restriction on a single identifier and adds the data field
to supply the different data.

One use case for this is to support card payments with different
modifiers for credit vs. debit cards.