Remove must in complete()
The complete method description indicated that the method must be called after the user has accepted the payment request.  This use of "must" was a requirement on the web page, and as an RFC 2119 term was misleading.  

Also, there is a SHOULD requirement on user agents in the paragraph, but the SHOULD was not in upper case so it was not called out as an assertion.