Add `enableLegacyCheckout` option to PaymentOptions.
Currently, when a payer cannot or does not want to choose one of
the payment methods provided in the PaymentRequest, they must
cancel it. However, it could be that they can still complete the
payment on the payee website via a custom checkout experience
(and their shipping information could still be propagated).
Requiring the payer to cancel the PaymentRequest in these
circumstances is less than graceful.

This patch adds an option to tell the user agent to provide an
option (on the user interface) to complete checkout on the payee's
website if they can't or do not want to select a payment method.
The payer can make this selection when necessary instead of
canceling the PaymentRequest.