Suggested emphasis of privacy protections
Perhaps my mistake, but I missed exactly how details of the the purchase (beyond payment scope) were protected from being sent to the payment processor.

@marcoscaceres helped make it clearer https://twitter.com/mrmarkrichards/status/909707142076227584 and suggested I raise a request here.

It'd be great to to ensure it is very clear that purchase item details, same as shipping details, shouldn't be shared with third parties (only the site you are paying would know).

Hopefully, this will reduce the likes of me querying it again. I think the wording suggested achieves the goal, but I'm not used to contributing to specs so feel free to criticise.